refactor: Split up MultiCompilerHost
The logic of the MultiCompilerHost almost only operated on one
CompilerHost plus respective CompilerOptions and caches, so it makes
sense to encapsulate this logic for one instance of these properties,
instead of requiring each of these properities to be a Record and having
to index them all the time.

The resulting code is less repetitive and more readable.